oz-backend-django/pyproject.toml
[tool.poetry]
name = "oz-backend-django"
version = "0.1.0"
description = "oz backend django"
authors = ["KangJenogHo1 <rkdwjdgh04@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
